scraper/requirements.txt
pika==1.2.0
beautifulsoup4==4.9.3
requests==2.25.1
python-dotenv